Move partial types
