+++ apis/Google.Cloud.Bigtable.V2/Google.Cloud.Bigtable.V2/BigtableClientPartial.cs
@@ -858,99 +858,5 @@ namespace Google.Cloud.Bigtable.V2
                     TableNameAsTableName = GaxPreconditions.CheckNotNull(tableName, nameof(tableName))
                 },
                 callSettings);
-
-        public partial class ReadRowsStream
-        {
-            private RowAsyncEnumerable _rowEnumerable;
-
-            /// <summary>
-            /// Returns an asynchronous sequence of rows from this set of results.
-            /// </summary>
-            /// <returns>An asynchronous sequence of rows from this set of results.</returns>
-            public IAsyncEnumerable<Row> AsAsyncEnumerable() =>
-                _rowEnumerable ?? (_rowEnumerable = new RowAsyncEnumerable(this));
-
-            private class RowAsyncEnumerable : IAsyncEnumerable<Row>
-            {
-                private int _enumeratorCount;
-                private ReadRowsStream _stream;
-
-                public RowAsyncEnumerable(ReadRowsStream stream)
-                {
-                    _stream = stream;
-                }
-
-                public IAsyncEnumerator<Row> GetEnumerator()
-                {
-                    if (Interlocked.CompareExchange(ref _enumeratorCount, 1, 0) == 1)
-                    {
-                        throw new InvalidOperationException(
-                            $"The result from {nameof(ReadRowsStream)}.{nameof(ReadRowsStream.AsAsyncEnumerable)} can only be iterated once");
-                    }
-
-                    return new RowAsyncEnumerator(_stream);
-                }
-            }
-        }
-
-        public partial class SampleRowKeysStream
-        {
-            /// <summary>
-            /// Gets all responses from the stream.
-            /// </summary>
-            /// <returns>The collection of all row key samples.</returns>
-            public IList<SampleRowKeysResponse> ToList() =>
-                Task.Run(() => ToListAsync()).ResultWithUnwrappedExceptions();
-
-            /// <summary>
-            /// Gets all responses from the stream asynchronously.
-            /// </summary>
-            /// <param name="cancellationToken">The token to monitor for cancellation requests.</param>
-            /// <returns>The collection of all row key samples.</returns>
-            public async Task<IList<SampleRowKeysResponse>> ToListAsync(
-                CancellationToken cancellationToken = default)
-            {
-                var responses = new List<SampleRowKeysResponse>();
-                while (await ResponseStream.MoveNext(cancellationToken).ConfigureAwait(false))
-                {
-                    responses.Add(ResponseStream.Current);
-                }
-                return responses;
-            }
-        }
-    }
-
-    public partial class BigtableClientImpl
-    {
-        private const string ResourcePrefixHeader = "google-cloud-resource-prefix";
-
-        partial void Modify_ReadRowsRequest(ref ReadRowsRequest request, ref CallSettings settings) =>
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-
-        partial void Modify_SampleRowKeysRequest(ref SampleRowKeysRequest request, ref CallSettings settings) =>
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-
-        partial void Modify_MutateRowRequest(ref MutateRowRequest request, ref CallSettings settings)
-        {
-            GaxPreconditions.CheckState(request.IsIdempotent(), "Non-idempotent MutateRow requests are not allowed. Specify a version with all SetCell mutations.");
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-        }
-
-        partial void Modify_MutateRowsRequest(ref MutateRowsRequest request, ref CallSettings settings)
-        {
-            GaxPreconditions.CheckState(request.IsIdempotent(), "Non-idempotent MutateRows requests are not allowed. Specify a version with all SetCell mutations.");
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-        }
-
-        partial void Modify_CheckAndMutateRowRequest(ref CheckAndMutateRowRequest request, ref CallSettings settings) =>
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-
-        partial void Modify_ReadModifyWriteRowRequest(ref ReadModifyWriteRowRequest request, ref CallSettings settings) =>
-            ApplyResourcePrefixHeader(ref settings, request.TableName);
-
-        private static void ApplyResourcePrefixHeader(ref CallSettings settings, string resource)
-        {
-            settings = settings.WithHeader(ResourcePrefixHeader, resource);
-        }
     }
 }
